+++ nodes/src/nodes/tool_firecrawl/IInstance.py
@@ -24,53 +24,25 @@
 # =============================================================================
 
 """
-Firecrawl tool node - global (shared) state.
+Firecrawl tool node instance.
 
-Reads the Firecrawl API key from config, creates a FirecrawlApp client,
-and builds a FirecrawlDriver that implements the ToolsBase interface.
+Delegates tool invoke operations to the FirecrawlDriver.
 """
 
 from __future__ import annotations
 
-from ai.common.config import Config
-from rocketlib import IGlobalBase, OPEN_MODE, warning
+from typing import Any
 
-from firecrawl import FirecrawlApp
+from rocketlib import IInstanceBase
 
-from .firecrawl_driver import FirecrawlDriver
+from .IGlobal import IGlobal
 
 
-class IGlobal(IGlobalBase):
-    """Global state for firecrawl_tools."""
+class IInstance(IInstanceBase):
+    IGlobal: IGlobal
 
-    driver: FirecrawlDriver | None = None
-
-    def beginGlobal(self) -> None:
-        if self.IEndpoint.endpoint.openMode == OPEN_MODE.CONFIG:
-            return
-
-        cfg = Config.getNodeConfig(self.glb.logicalType, self.glb.connConfig)
-
-        apikey = str((cfg.get('apikey') or '')).strip()
-
-        if not apikey:
-            raise Exception('firecrawl_tools: apikey is required')
-
-        try:
-            app = FirecrawlApp(api_key=apikey)
-            self.driver = FirecrawlDriver(server_name='firecrawl', app=app)
-        except Exception as e:
-            warning(str(e))
-            raise
-
-    def validateConfig(self) -> None:
-        try:
-            cfg = Config.getNodeConfig(self.glb.logicalType, self.glb.connConfig)
-            apikey = str((cfg.get('apikey') or '')).strip()
-            if not apikey:
-                warning('apikey is required')
-        except Exception as e:
-            warning(str(e))
-
-    def endGlobal(self) -> None:
-        self.driver = None
+    def invoke(self, param: Any) -> Any:  # noqa: ANN401
+        driver = getattr(self.IGlobal, 'driver', None)
+        if driver is None:
+            raise RuntimeError('tool_firecrawl: driver not initialized')
+        return driver.handle_invoke(param)
